Ensure news and research have different ids
Id is used as key in [...slug].page page and hence if news and research
share the same id, React doesn't rerender the articles component when a
user switches between news and research pages. This affects when coming
from pages 2 of say news and you want to go to research which should
start from page one.